+++ src/rath/flow/workflow/__init__.py
@@ -1,0 +1,5 @@
+"""Workflow API."""
+
+from rath.flow.workflow.module import SingleAgent, Workflow
+
+__all__ = ["Workflow", "SingleAgent"]